- Changed the scale and shift parameters for a layered texture   from int to float. Also added a logical change. CS_FX_TILING   should now be used when you want the layer texture to tile on   the base texture. This used to be default but it isn't any more.   Note that the OpenGL renderer currently only supports the tiling   option so the presence of CS_FX_TILING or not will not have   an effect currently. - Added support for TILING() keyword in the map loader and all loader   plugins.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7455 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csengine/material.h
@@ -96,7 +96,7 @@ public:
 
   /// Add a texture layer (currently only one supported).
   void AddTextureLayer (iTextureWrapper* txtwrap, UInt mode,
-      	int uscale, int vscale, int ushift, int vshift);
+      	float uscale, float vscale, float ushift, float vshift);
 
   //--------------------- iMaterial implementation ---------------------
